+++ new_import_from_here/amaderbari/src/main/java/sdp/ratajo/amaderbari/dto/flatpack/Flat.java
@@ -14,10 +14,10 @@ import org.hibernate.validator.constraints.URL;
 
 public class Flat {
 	
-	private long flatId;
-	private long addressId;
-	private long extraFlatDataId;
-	private long imageId;
+	private Integer flatId;
+	private Integer addressId;
+	private Integer extraFlatDataId;
+	private Integer imageId;
 	
 	@Size(min=3, max=50, message="The length of data must be between 3 and 50")
 	@Email(message="not valid email")
@@ -39,7 +39,7 @@ public class Flat {
 	
 	public Flat(){}
 
-	public Flat(long flatId, long addressId, long extraFlatDataId, long imageId,
+	public Flat(Integer flatId, Integer addressId, Integer extraFlatDataId, Integer imageId,
 			String ownerEmail, String renterEmail, String mapUrl,
 			String squareFoot, double rent) {
 		super();
@@ -54,22 +54,22 @@ public class Flat {
 		this.rent = rent;
 	}
 
-	public long getFlatId() {
+	public Integer getFlatId() {
 		return flatId;
 	}
 
 
-	public void setFlatId(long flatId) {
+	public void setFlatId(Integer flatId) {
 		this.flatId = flatId;
 	}
 
 
-	public long getAddressId() {
+	public Integer getAddressId() {
 		return addressId;
 	}
 
 
-	public void setAddressId(long addressId) {
+	public void setAddressId(Integer addressId) {
 		this.addressId = addressId;
 	}
 
@@ -124,22 +124,22 @@ public class Flat {
 	}
 
 
-	public long getExtraFlatDataId() {
+	public Integer getExtraFlatDataId() {
 		return extraFlatDataId;
 	}
 
 
-	public void setExtraFlatDataId(long extraFlatDataId) {
+	public void setExtraFlatDataId(Integer extraFlatDataId) {
 		this.extraFlatDataId = extraFlatDataId;
 	}
 
 
-	public long getImageId() {
+	public Integer getImageId() {
 		return imageId;
 	}
 
 
-	public void setImageId(long imageId) {
+	public void setImageId(Integer imageId) {
 		this.imageId = imageId;
 	}
